Fix confirmation dialog not popping the context and make it a reusable widget ConfirmDialog being a widget was popping its own context and not the settings page one
